comparing prediction quality
[am I doing it correctly?]

Currently the prediction quality sucks; I think this may be because the line I
crafted with theta fits the unnormalized data nicely, but not the normalized one
where the points are in a different place. I can try *not* to feature-scale
this data (as it's artificial), or alternatively find a line that's good for
the normalized data. Alternatively maybe I can scale the line somehow?