Start working on hinge loss
I want to create a reasonably vectorized version. But for that I need detailed
tests. So the idea is to add a trivial unvectorized version in
binomial_logistic_regression_test and write unit tests for it. Then I can
comprehensively test the vectorized implementation vs. the "golden" simple one.

Another idea is to implement a gradient check.